GenericArguments consults the wrong state when tracking modified argument descriptors and mapped arguments https://bugs.webkit.org/show_bug.cgi?id=174411 <rdar://problem/31696186>
Reviewed by Mark Lam.

JSTests:

* stress/generic-arguments-correct-delete-behavior.js: Added.
(assert):
(makeTest):

Source/JavaScriptCore:

The code for deleting an argument was incorrectly referencing state
when it decided if it should unmap or mark a property as having its
descriptor modified. This patch fixes the bug where if we delete a
property, we would sometimes not unmap an argument when deleting it.

* runtime/GenericArgumentsInlines.h:
(JSC::GenericArguments<Type>::getOwnPropertySlot):
(JSC::GenericArguments<Type>::getOwnPropertySlotByIndex):
(JSC::GenericArguments<Type>::deleteProperty):
(JSC::GenericArguments<Type>::deletePropertyByIndex):


Identifier: 191264@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@219433 268f45cc-cd09-0410-ab3c-d52691b4dbfc