Possible crash under NetworkSocketStream::didFailSocketStream() https://bugs.webkit.org/show_bug.cgi?id=174526 <rdar://problem/32831441>
Reviewed by Brent Fulgham.

Source/WebCore:

Call m_client.didFailSocketStream() asynchronously in the constructor as our
caller (the client) is also being initialized at this point.

* platform/network/cf/SocketStreamHandleImplCFNet.cpp:
(WebCore::SocketStreamHandleImpl::SocketStreamHandleImpl):

Source/WebKit:

For robustness, initialize the SocketStreamHandleImpl after the other
data members. We are passing |this| to the SocketStreamHandleImpl when
constructing it so it is unsafe to have uninitialized data members
at this point.

* NetworkProcess/NetworkSocketStream.cpp:
(WebKit::NetworkSocketStream::NetworkSocketStream):
* NetworkProcess/NetworkSocketStream.h:


Identifier: 191347@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@219525 268f45cc-cd09-0410-ab3c-d52691b4dbfc