Web Automation: link and partial link queries don't work if the link contains formatting tags https://bugs.webkit.org/show_bug.cgi?id=174498
Reviewed by Brian Burg.

So, for example, if we find links with text "Foo" and there's a link like <a href=""><bA>Foo</b></a> we fail
with no such element error. This causes test test_Link_With_Formatting_Tags to fail.

* UIProcess/Automation/atoms/FindNodes.js:
(switch): Use descendant-or-self::text() instead of just text() in the links xpath expressions.

Identifier: 191377@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@219555 268f45cc-cd09-0410-ab3c-d52691b4dbfc