Butterfly storage need not be initialized for indexing type Undecided. https://bugs.webkit.org/show_bug.cgi?id=174516
Reviewed by Saam Barati.

While it's not incorrect to initialize the butterfly storage when the
indexingType is Undecided, it is inefficient as we'll end up initializing
it again later when we convert the storage to a different indexingType.
Some of our code already skips initializing Undecided butterflies.
This patch makes it the consistent behavior everywhere.

* dfg/DFGSpeculativeJIT.cpp:
(JSC::DFG::SpeculativeJIT::emitAllocateRawObject):
* runtime/JSArray.cpp:
(JSC::JSArray::tryCreateUninitializedRestricted):
* runtime/JSArray.h:
(JSC::JSArray::tryCreate):
* runtime/JSObject.cpp:
(JSC::JSObject::ensureLengthSlow):


Identifier: 191447@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@219636 268f45cc-cd09-0410-ab3c-d52691b4dbfc