AX: Web page reloaded when a node is labelling multiple childnodes https://bugs.webkit.org/show_bug.cgi?id=174655
Reviewed by Chris Fleizach.

Source/WebCore:

When we are asking for the aria-labelledby attribute of a node and its
sibling is also labelled by the same node, we get into an infinite loop
in textUnderElement since we only ignore one child. Added checks for
siblings to avoid such loop.

Test: accessibility/mac/aria-labelledby-multiple-child-crash.html

* accessibility/AccessibilityNodeObject.cpp:
(WebCore::AccessibilityNodeObject::textUnderElement):

LayoutTests:

* accessibility/mac/aria-labelledby-multiple-child-crash-expected.txt: Added.
* accessibility/mac/aria-labelledby-multiple-child-crash.html: Added.


Identifier: 191472@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@219661 268f45cc-cd09-0410-ab3c-d52691b4dbfc