[GTK][WPE] Remove useless conditional in IconDatabase::setIconDataForIconURL https://bugs.webkit.org/show_bug.cgi?id=174792
Patch by Michael Catanzaro <mcatanzaro@igalia.com> on 2017-07-24
Reviewed by Carlos Garcia Campos.

We have in IconDatabase::setIconDataForIconURL some code in a conditional to check if it's
not being executed on the icon sync thread. But at the top of the function is an assertion
to ensure the code is never executed on the icon sync thread. The condition is therefore
useless and the code should be executed unconditionally.

* UIProcess/API/glib/IconDatabase.cpp:
(WebKit::IconDatabase::setIconDataForIconURL):

Identifier: 191626@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@219862 268f45cc-cd09-0410-ab3c-d52691b4dbfc